fix: #41 Change wording in README
Changed the heading to show the word exclusive as it's first indicator to it being an exclusive project.

Added a small note heading with information underneath linking to scrimba bootcamp, they wish to take part.

Changed the see it live link as well as some emotes that were not showing up on my system locally.