feat/fix: #56 fix color & comment code & add feat
I adjusted the color to a css variable and placed the color
within the base.css file so that colors are not all over the
place, i did credit the owner.
I commented the code to what I thought was happening, to help other students.
I also adjusted and added to the component to include a message in the
hero heading as that seemed more reasonable as to not have another greeting in the middle of the page.